requirements.txt
accelerate==1.1.1
c4d_perftracker_collector==1.4.0
configs==3.0.3
datasets==1.6.2
einops==0.8.1
fastmri==0.3.0
flash_attn==2.7.0.post2
flash_linear_attention==0.2.2
h5py==3.13.0
lm_eval==0.4.8
lmdb==1.6.2
mmengine==0.10.5
numpy==2.3.2
pandas==2.3.1
PyYAML==6.0.2
safetensors==0.6.2
torch==2.5.0
tqdm==4.67.0
transformers==4.52.4
wandb==0.21.1
xtuner==0.2.0

